+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Add the return field `accessibleTrackId` in EmmPolicyApi - `getResellerEmmPolicy`,`getMerchantEmmPolicy`,`createResellerEmmPolicy`,`createMerchantEmmPolicy` API
 * Add parameter query `serialNo` 、`pidList` and  return in TerminalApkApi-`searchTerminalApk` API
 * Add parameter query `serialNo` and  return in TerminalRkiApi-`searchPushRkiTasks` API
+* Add TerminalApkApi - `deleteTerminalApk` API
+* Add TerminalFirmwareApi - `deleteTerminalFirmware` API
+* Add TerminalRkiApi - `deleteTerminalRki` API
 
 
 ## 9.7.1